% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{badge_cran_downloads}
\alias{badge_cran_downloads}
\title{Add a badge for downloads from CRAN}
\usage{
badge_cran_downloads(packagename, period = NULL)
}
\arguments{
\item{packagename}{name of the package}

\item{period}{defaults to month, other options are last-week, last-day, grand-total}
}
\value{
markdown
}
\description{
Display the number of downloads from CRAN. Use
last-week, last-day, or grand-total, defaults to montly.
}
\examples{
badge_cran_downloads("dplyr", period = "last-week")
}
