% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{badge_cran_version_ago}
\alias{badge_cran_version_ago}
\title{CRAN version and release in time}
\usage{
badge_cran_version_ago(packagename)
}
\arguments{
\item{packagename}{the name of your package}
}
\value{
markdown
}
\description{
See \url{https://r-pkg.org/services#badges}
One of the metacran badges.
}
\examples{
badge_cran_version_ago("dplyr")
}
\seealso{
Other badges: \code{\link{badge_codecov}},
  \code{\link{badge_cran_ago}},
  \code{\link{badge_cran_date}},
  \code{\link{badge_cran_version_release}},
  \code{\link{badge_cran}},
  \code{\link{badge_github_fork}},
  \code{\link{badge_github_star}},
  \code{\link{badge_last_change_static}},
  \code{\link{badge_last_change}},
  \code{\link{badge_license}},
  \code{\link{badge_minimal_r_version}},
  \code{\link{badge_packageversion}},
  \code{\link{badge_projectstatus}},
  \code{\link{badge_rank}},
  \code{\link{badge_rdocumentation}},
  \code{\link{badge_thanks_md}}, \code{\link{badge_travis}}
}
