% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{badge_cran_version_release}
\alias{badge_cran_version_release}
\title{CRAN version and date of release}
\usage{
badge_cran_version_release(packagename)
}
\arguments{
\item{packagename}{the name of your package}
}
\value{
markdown
}
\description{
CRAN version and date of release
}
\examples{
badge_cran_version_release("dplyr")
}
\seealso{
Other badges: \code{\link{badge_codecov}},
  \code{\link{badge_cran_ago}},
  \code{\link{badge_cran_date}},
  \code{\link{badge_cran_version_ago}},
  \code{\link{badge_cran}},
  \code{\link{badge_github_fork}},
  \code{\link{badge_github_star}},
  \code{\link{badge_last_change_static}},
  \code{\link{badge_last_change}},
  \code{\link{badge_license}},
  \code{\link{badge_minimal_r_version}},
  \code{\link{badge_packageversion}},
  \code{\link{badge_projectstatus}},
  \code{\link{badge_rank}},
  \code{\link{badge_rdocumentation}},
  \code{\link{badge_travis}}
}
