% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{badge_last_change}
\alias{badge_last_change}
\title{creates last-change badge}
\usage{
badge_last_change(location = ".")
}
\arguments{
\item{location}{defaults to working directory}
}
\description{
Will add a badge containing the current date that changes 
on every reknitting. This is a simple pasting of r code.
}
\seealso{
Other badges: \code{\link{badge_codecov}},
  \code{\link{badge_cran}},
  \code{\link{badge_github_fork}},
  \code{\link{badge_github_star}},
  \code{\link{badge_last_change_static}},
  \code{\link{badge_licence}},
  \code{\link{badge_minimal_r_version}},
  \code{\link{badge_packageversion}},
  \code{\link{badge_projectstatus}},
  \code{\link{badge_rdocumentation}},
  \code{\link{badge_travis}}, \code{\link{licbadgebuilder}}
}
