% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{badge_licence}
\alias{badge_licence}
\title{Create a licencebadge}
\usage{
badge_licence(licence = NULL, location = ".")
}
\arguments{
\item{licence}{License, for example `GPL-3`, `MIT`, etc. Alternatively, `search`.}

\item{location}{defaults to current directory}
}
\value{
markdown
}
\description{
when `licence="search"`` this function wil look in your
DESCRIPTION file and search for the "Licence:" part. 
If it matches GPL or MIT a custum badge will be created. 
If it does not match, a general badge will be created with the name of the
licence in grey.
}
\seealso{
Other badges: \code{\link{badge_codecov}},
  \code{\link{badge_cran}},
  \code{\link{badge_github_fork}},
  \code{\link{badge_github_star}},
  \code{\link{badge_last_change_static}},
  \code{\link{badge_last_change}},
  \code{\link{badge_minimal_r_version}},
  \code{\link{badge_packageversion}},
  \code{\link{badge_projectstatus}},
  \code{\link{badge_rdocumentation}},
  \code{\link{badge_travis}}, \code{\link{licbadgebuilder}}
}
