% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{badge_license}
\alias{badge_license}
\title{Create a licensebadge}
\usage{
badge_license(license = NULL, location = ".")
}
\arguments{
\item{license}{License, for example \code{GPL-3}, \code{MIT}, etc. Alternatively, leaving it empty will search.}

\item{location}{defaults to current directory}
}
\value{
markdown
}
\description{
when `license= NULL' this function will look in your
DESCRIPTION file and search for the "license:" part.
If it matches GPL or MIT a custom badge will be created.
If it does not match, a general badge will be created with the name of the
license in grey.
}
\seealso{
Other badges: \code{\link{badge_codecov}},
  \code{\link{badge_cran_ago}},
  \code{\link{badge_cran_date}},
  \code{\link{badge_cran_version_ago}},
  \code{\link{badge_cran_version_release}},
  \code{\link{badge_cran}},
  \code{\link{badge_github_fork}},
  \code{\link{badge_github_star}},
  \code{\link{badge_last_change_static}},
  \code{\link{badge_last_change}},
  \code{\link{badge_minimal_r_version}},
  \code{\link{badge_packageversion}},
  \code{\link{badge_projectstatus}},
  \code{\link{badge_rank}},
  \code{\link{badge_rdocumentation}},
  \code{\link{badge_thanks_md}}, \code{\link{badge_travis}}
}
