% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{badge_lifecycle}
\alias{badge_lifecycle}
\title{Add a lifecycle badge}
\source{
\url{www.tidyverse.org/lifecycle}
}
\usage{
badge_lifecycle(lifecycle = "experimental")
}
\description{
Lifecycle denotations were introduced by the tidyverse
team in 2018. Because project status was not a perfect fit.
I think it is a real professional attitude, the badges
communicate to your
users what they can expect.
}
