% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{badge_minimal_r_version}
\alias{badge_minimal_r_version}
\title{Display the minimal R version}
\usage{
badge_minimal_r_version(chunk = TRUE)
}
\value{
markdown
}
\description{
Display the minimal R version
}
\examples{
badge_minimal_r_version()
}
\seealso{
Other badges: \code{\link{badge_codecov}},
  \code{\link{badge_cran}},
  \code{\link{badge_github_fork}},
  \code{\link{badge_github_star}},
  \code{\link{badge_last_change_static}},
  \code{\link{badge_last_change}},
  \code{\link{badge_licence}},
  \code{\link{badge_packageversion}},
  \code{\link{badge_projectstatus}},
  \code{\link{badge_rdocumentation}},
  \code{\link{badge_travis}}, \code{\link{licbadgebuilder}}
}
