% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{badge_packageversion}
\alias{badge_packageversion}
\title{Place a badge with the version of your package.}
\usage{
badge_packageversion(chunk = TRUE)
}
\arguments{
\item{chunk}{this argument places a rmarkdown chunk}
}
\value{
markdown to put into readme.rmd
}
\description{
Place a badge with the version of your package which is automatically read from
your description file.
}
\examples{
badge_packageversion()
}
\seealso{
Other badges: \code{\link{badge_codecov}},
  \code{\link{badge_cran}},
  \code{\link{badge_github_fork}},
  \code{\link{badge_github_star}},
  \code{\link{badge_last_change_static}},
  \code{\link{badge_last_change}},
  \code{\link{badge_licence}},
  \code{\link{badge_minimal_r_version}},
  \code{\link{badge_projectstatus}},
  \code{\link{badge_rdocumentation}},
  \code{\link{badge_travis}}, \code{\link{licbadgebuilder}}
}
