% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{badge_projectstatus}
\alias{badge_projectstatus}
\title{Add project status badge}
\usage{
badge_projectstatus(status = NULL)
}
\arguments{
\item{status}{one of concept, wip, suspended, abandoned, active, inactive or unsupported}
}
\value{
text to put into rmd
}
\description{
Add project status badge
}
\examples{
 
badge_projectstatus("unsupported")
}
\seealso{
Other badges: \code{\link{badge_codecov}},
  \code{\link{badge_cran}},
  \code{\link{badge_github_fork}},
  \code{\link{badge_github_star}},
  \code{\link{badge_last_change_static}},
  \code{\link{badge_last_change}},
  \code{\link{badge_licence}},
  \code{\link{badge_minimal_r_version}},
  \code{\link{badge_packageversion}},
  \code{\link{badge_rdocumentation}},
  \code{\link{badge_travis}}, \code{\link{licbadgebuilder}}
}
