% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{badge_thanks_md}
\alias{badge_thanks_md}
\title{Add thanks badge}
\source{
\url{https://github.com/paulmolluzzo/thanks-md}
}
\usage{
badge_thanks_md(add_file = TRUE)
}
\arguments{
\item{add_file}{Should the badge add the THANKS.md file to your project?}
}
\value{
markdown
}
\description{
Add a thanks badge and THANKS.md file to your project.
}
\seealso{
Other badges: \code{\link{badge_codecov}},
  \code{\link{badge_cran_ago}},
  \code{\link{badge_cran_date}},
  \code{\link{badge_cran_version_ago}},
  \code{\link{badge_cran_version_release}},
  \code{\link{badge_cran}},
  \code{\link{badge_github_fork}},
  \code{\link{badge_github_star}},
  \code{\link{badge_last_change_static}},
  \code{\link{badge_last_change}},
  \code{\link{badge_license}},
  \code{\link{badge_minimal_r_version}},
  \code{\link{badge_packageversion}},
  \code{\link{badge_projectstatus}},
  \code{\link{badge_rank}},
  \code{\link{badge_rdocumentation}},
  \code{\link{badge_travis}}
}
\concept{badges}
