% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badgecreatr.R
\docType{package}
\name{badgecreatr}
\alias{badgecreatr}
\alias{badgecreatr-package}
\title{badgecreatr}
\description{
Create badges in your readme file on top of a github repository.
This package creates badges for a wide range of functions.
}
\details{
You can place a fixed set of functions in your readme, add badges only
or add interactive versions with embedded R-code
that changes when you re-knit your badges.
}
