% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{badgepaste}
\alias{badgepaste}
\title{return markdown with image and link.}
\usage{
badgepaste(imagelink, referlink, name = "license")
}
\arguments{
\item{imagelink}{link to image file}

\item{referlink}{link to where to send to on click}

\item{name}{what to call the button}
}
\value{
markdown
}
\description{
for internal use. used within licensebuilder
}
