% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{cranbadge}
\alias{cranbadge}
\title{Add a badge for cran}
\usage{
cranbadge(packagename)
}
\arguments{
\item{packagename}{the name of your package}
}
\value{
markdown to put into readme
}
\description{
Add a badge for cran
}
\examples{
cranbadge("badgecreatr")
}

