% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_information.R
\name{get_remote_reponame_username}
\alias{get_remote_reponame_username}
\title{Return the name of the repo and and username.}
\usage{
get_remote_reponame_username()
}
\description{
This information can be fed into the badeplacers.
}
