% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{github_credentials_helper}
\alias{github_credentials_helper}
\title{Returns github accountname, repo name and current branch.}
\usage{
github_credentials_helper(ghaccount = NULL, ghrepo = NULL, branch = NULL,
  location = ".")
}
\arguments{
\item{ghaccount}{Github account name}

\item{ghrepo}{Github repository name}

\item{branch}{branch name such as master, develop etc}

\item{location}{defaults to current location}
}
\description{
Returns github accountname, repo name and current branch.
}
