% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_recognition.R
\name{githubcredentials}
\alias{githubcredentials}
\title{Give or search for github name and repo.
If you don't put anything here, leaving everything empty,
this function will check the git repository for information.}
\usage{
githubcredentials(account = NULL, repo = NULL, branch = NULL)
}
\arguments{
\item{account}{accountname}

\item{repo}{repositoryname}

\item{branch}{master, develop etc.}
}
\value{
markdown text to be placed in readme
}
\description{
Give or search for github name and repo.
If you don't put anything here, leaving everything empty,
this function will check the git repository for information.
}
