% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{last_change_badge2}
\alias{last_change_badge2}
\title{creates last-change badge}
\usage{
last_change_badge2(location = ".")
}
\arguments{
\item{location}{defaults to working directory}
}
\description{
Will add current day to the repo.
}
