% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{last_change_badge_static}
\alias{last_change_badge_static}
\title{creates last-change badge}
\usage{
last_change_badge_static(location = ".")
}
\arguments{
\item{location}{defaults to working directory}
}
\description{
Will add current day to the repo.
}
\seealso{
Other badges: \code{\link{codecovbadge}},
  \code{\link{cranbadge}}, \code{\link{last_change_badge}},
  \code{\link{licbadgebuilder}},
  \code{\link{minimal_r_version_badge}},
  \code{\link{packageversionbadge}},
  \code{\link{projectstatusbadge}},
  \code{\link{travisbadge}}
}
