% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{licbadgebuilder}
\alias{licbadgebuilder}
\title{Add license badge}
\usage{
licbadgebuilder(licensetype)
}
\arguments{
\item{licensetype}{one of GPL-3, GPL-2, MIT, or CC0.}
}
\value{
markdown
}
\description{
Add license badge
}
\examples{
badgecreatr:::licbadgebuilder("GPL-3")
}
