% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{licencebadge}
\alias{licencebadge}
\title{Create a licencebadge}
\usage{
licencebadge(licence = "search", location = ".")
}
\arguments{
\item{licence}{License, for example `GPL-3`, `MIT`, etc. Alternatively, `search`.}
}
\value{
markdown
}
\description{
when `licence="search"`` this function wil look in your
DESCRIPTION file and search for the "Licence:" part. 
If it matches GPL or MIT a custum badge will be created. 
If it does not match, a general badge will be created with the name of the
licence in grey.
}
