% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{minimal_r_version_badge}
\alias{minimal_r_version_badge}
\title{Display the minimal R version}
\usage{
minimal_r_version_badge()
}
\value{
markdown
}
\description{
Display the minimal R version
}
\examples{
minimal_r_version_badge()
}

