% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{packageversionbadge}
\alias{packageversionbadge}
\title{place a badge with the version of your package}
\usage{
packageversionbadge(packageversionnumber)
}
\arguments{
\item{packageversionnumber}{give by hand or let it search in description file.}
}
\value{
markdown to put into readme.rmd
}
\description{
place a badge with the version of your package
}
\examples{
packageversionbadge("0.0.2")
}

