% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{projectstatusbadge}
\alias{projectstatusbadge}
\title{Add project status badge}
\usage{
projectstatusbadge(status = NULL)
}
\arguments{
\item{status}{one of concept, wip, suspended, abandoned, active, inactive or unsupported}
}
\value{
text to put into rmd
}
\description{
Add project status badge
}
\examples{
 
projectstatusbadge("unsupported")
}
\seealso{
Other badges: \code{\link{codecovbadge}},
  \code{\link{cranbadge}},
  \code{\link{last_change_badge_static}},
  \code{\link{last_change_badge}},
  \code{\link{licbadgebuilder}},
  \code{\link{minimal_r_version_badge}},
  \code{\link{packageversionbadge}},
  \code{\link{travisbadge}}
}
