% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{rdocumentation_badge}
\alias{rdocumentation_badge}
\title{R documentation badge}
\usage{
rdocumentation_badge(packagename)
}
\arguments{
\item{packagename}{the name of your package}
}
\description{
Add a documentation badge, from DataCamp.
}
\examples{
rdocumentation_badge("dplyr")
}
