% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_information.R
\name{return_remotes}
\alias{return_remotes}
\title{Return a list of remotes associated with this repo}
\usage{
return_remotes(debug = FALSE)
}
\description{
a function that checks the remote branches 
and returns their names.
}
