% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_information.R
\name{search_git}
\alias{search_git}
\title{Return a list of remotes associated with this repo}
\usage{
search_git(location = ".")
}
\description{
a function that checks the remote branches 
and returns their names.
}
