% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{travisbadge}
\alias{travisbadge}
\title{Travisbadge creates travis badge.}
\usage{
travisbadge(createbadge = TRUE, ghaccount, ghrepo, branch)
}
\arguments{
\item{createbadge}{a TRUE or FALSE for checking.}

\item{ghaccount}{githubaccountname}

\item{ghrepo}{githubrepositoryname}

\item{branch}{master, develop etc}
}
\value{
link to travis image
}
\description{
Travisbadge creates travis badge.
}
\examples{
travisbadge(createbadge = TRUE, ghaccount = "johntest", ghrepo = "badgecreatr", branch = "master")
}

