% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{badge_codecov}
\alias{badge_codecov}
\title{CodeCoverage ' ' Adds a code cov badge}
\usage{
badge_codecov(ghaccount = NULL, ghrepo = NULL, branch = NULL,
  location = ".")
}
\arguments{
\item{ghaccount}{githubaccountname}

\item{ghrepo}{githubrepositoryname}

\item{branch}{master, develop etc}

\item{location}{defaults to current location}
}
\description{
Adds codecov badge
}
\examples{
badge_codecov(ghaccount = "johntest", ghrepo = "yourreponame", branch = "master")
}
\seealso{
Other badges: \code{\link{badge_cran}},
  \code{\link{badge_github_fork}},
  \code{\link{badge_github_star}},
  \code{\link{badge_last_change_static}},
  \code{\link{badge_last_change}},
  \code{\link{badge_licence}},
  \code{\link{badge_minimal_r_version}},
  \code{\link{badge_packageversion}},
  \code{\link{badge_projectstatus}},
  \code{\link{badge_rdocumentation}},
  \code{\link{badge_travis}}, \code{\link{licbadgebuilder}}
}
