% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{badge_cran}
\alias{badge_cran}
\title{Add a badge for cran}
\usage{
badge_cran(packagename)
}
\arguments{
\item{packagename}{the name of your package}
}
\value{
markdown to put into readme
}
\description{
Add a badge for cran
}
\examples{
badge_cran("dplyr")
}
\seealso{
Other badges: \code{\link{badge_codecov}},
  \code{\link{badge_github_fork}},
  \code{\link{badge_github_star}},
  \code{\link{badge_last_change_static}},
  \code{\link{badge_last_change}},
  \code{\link{badge_licence}},
  \code{\link{badge_minimal_r_version}},
  \code{\link{badge_packageversion}},
  \code{\link{badge_projectstatus}},
  \code{\link{badge_rdocumentation}},
  \code{\link{badge_travis}}, \code{\link{licbadgebuilder}}
}
