% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{badge_last_change}
\alias{badge_last_change}
\title{Creates last-change badge}
\usage{
badge_last_change(location = ".")
}
\arguments{
\item{location}{defaults to working directory}
}
\value{
Rmarkdown
}
\description{
Will add a badge containing the current date that changes
on every reknitting. This is a simple pasting of r-code.
}
\seealso{
Other badges: \code{\link{badge_codecov}},
  \code{\link{badge_cran_ago}},
  \code{\link{badge_cran_date}},
  \code{\link{badge_cran_version_ago}},
  \code{\link{badge_cran_version_release}},
  \code{\link{badge_cran}},
  \code{\link{badge_github_fork}},
  \code{\link{badge_github_star}},
  \code{\link{badge_last_change_static}},
  \code{\link{badge_license}},
  \code{\link{badge_minimal_r_version}},
  \code{\link{badge_packageversion}},
  \code{\link{badge_projectstatus}},
  \code{\link{badge_rank}},
  \code{\link{badge_rdocumentation}},
  \code{\link{badge_thanks_md}}, \code{\link{badge_travis}}
}
