% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{badge_last_change_static}
\alias{badge_last_change_static}
\title{Creates last-change badge static}
\usage{
badge_last_change_static(date = NULL)
}
\arguments{
\item{date}{if NULL/empty current date. otherwise use yyyy-mm-dd format}
}
\value{
markdown
}
\description{
Will add current day to the repo. This function will
not change when you reknit the readme. If you want that you will
have to use the link{last_change_badge} function.
}
\seealso{
Other badges: \code{\link{badge_codecov}},
  \code{\link{badge_cran_ago}},
  \code{\link{badge_cran_date}},
  \code{\link{badge_cran_version_ago}},
  \code{\link{badge_cran_version_release}},
  \code{\link{badge_cran}},
  \code{\link{badge_github_fork}},
  \code{\link{badge_github_star}},
  \code{\link{badge_last_change}},
  \code{\link{badge_licence}},
  \code{\link{badge_minimal_r_version}},
  \code{\link{badge_packageversion}},
  \code{\link{badge_projectstatus}},
  \code{\link{badge_rank}},
  \code{\link{badge_rdocumentation}},
  \code{\link{badge_travis}}
}
