% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badgecreatr.R
\docType{package}
\name{badgecreatr}
\alias{badgecreatr}
\alias{badgecreatr-package}
\title{badgecreatr}
\description{
Create badges in your readme file 
with a simple function. 
Run badgecreatr once and the badges are created.
Run it again the badges are refreshed.
}
\details{
Other options include badges only and interactive versions
that change when you re-knit your badges.
}
