% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badgecreatr.R
\docType{package}
\name{badgecreatr}
\alias{badgecreatr}
\alias{badgecreatr-package}
\title{badgecreatr}
\description{
Create badges in your readme file 
with a simple function. Run this once and the badges are created.
}
