% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagechecks.R
\name{badgeplacer}
\alias{badgeplacer}
\title{Place badges inside readme file}
\usage{
badgeplacer(location = ".", status = "active", githubaccount = "search",
  githubrepo = "search", branch = "master", name = "README.Rmd")
}
\arguments{
\item{location}{folder where readme.rmd resides}

\item{status}{one of: "concept", "wip", "suspended", "abandoned", "active", "inactive",  or "unsupported"}

\item{githubaccount}{your githubname}

\item{githubrepo}{your githubrepositoryname}

\item{branch}{master, develop, etc.}
}
\value{
readme file with added badges
}
\description{
This is the main function to add badges to your readme.
}

