% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagechecks.R
\name{badgeplacer}
\alias{badgeplacer}
\title{Place badges inside readme file}
\usage{
badgeplacer(location = ".")
}
\value{
readme file with added badges
}
\description{
This is the main function to add badges to your readme.
}

