% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{codecovbadge}
\alias{codecovbadge}
\title{CodeCoverage ' ' Adds a code cov badge}
\usage{
codecovbadge(ghaccount, ghrepo, branch = "master")
}
\arguments{
\item{ghaccount}{your github account f.i. "rmhogervorst"}

\item{ghrepo}{the name of the repo f.i. "badgecreatr"}

\item{branch}{the branch, defaults to master}
}
\description{
adds codecov badge
}
\examples{
codecovbadge(ghaccount = "johntest", ghrepo = "yourreponame", branch = "master")
}
