% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{codecovbadge}
\alias{codecovbadge}
\title{CodeCoverage ' ' Adds a code cov badge}
\usage{
codecovbadge(ghaccount = "search", ghrepo = "search", branch = "master")
}
\arguments{
\item{ghaccount}{your github account f.i. "rmhogervorst",you can use "search"}

\item{ghrepo}{the name of the repo f.i. "badgecreatr",you can use "search"}

\item{branch}{the branch, defaults to master, you can use "search"}
}
\description{
Adds codecov badge
}
\examples{
codecovbadge(ghaccount = "johntest", ghrepo = "yourreponame", branch = "master")
}
\seealso{
Other badges: \code{\link{cranbadge}},
  \code{\link{last_change_badge_static}},
  \code{\link{last_change_badge}},
  \code{\link{licbadgebuilder}},
  \code{\link{minimal_r_version_badge}},
  \code{\link{packageversionbadge}},
  \code{\link{projectstatusbadge}},
  \code{\link{travisbadge}}
}
