% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{dynamic_badges_minimal}
\alias{dynamic_badges_minimal}
\title{Add dynamic content to readme}
\usage{
dynamic_badges_minimal(status = "concept", licence = "search",
  last_change = TRUE, minimal_r_version = TRUE, travisfile = NULL,
  codecov = NULL, location = ".")
}
\arguments{
\item{status}{one of concept, wip, suspended, abandoned, active, inactive or unsupported}

\item{licence}{one of GPL-3, GPL-2, MIT, or CC0.}

\item{last_change}{adding a last change badge? TRUE or FALSE}

\item{minimal_r_version}{adding minimal r version badge? TRUE or FALSE}

\item{travisfile}{add a travis statusbadge}

\item{codecov}{add a codecov badge}

\item{location}{defaults to current location}
}
\value{
text to put into rmd
}
\description{
This function returns markdown that you must place in the readme. This is
dynamic content. This function returns a RMarkdown piece that does some
calculation and a set of badges underneath.
}
\seealso{
Other scenarios: \code{\link{minimal_badges}}
}
