% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{findbadges}
\alias{findbadges}
\title{What badges are already found in the Readme document}
\usage{
findbadges(location = ".")
}
\arguments{
\item{location}{where should we search for readme.rmd?}
}
\value{
list of locations inside of readme
}
\description{
What badges are already found in the Readme document
}

