% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{findbadges}
\alias{findbadges}
\title{What badges are already found in the Readme document}
\usage{
findbadges(location = ".", name = "README.Rmd")
}
\arguments{
\item{location}{where should we search for readme.rmd?}

\item{name}{which file to place badges in defaults to README.Rmd}
}
\value{
list of locations inside of readme
}
\description{
What badges are already found in the Readme document
}

