% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_recognition.R
\name{githubcredentials}
\alias{githubcredentials}
\title{Give or search for github name and repo.}
\usage{
githubcredentials(account = "search", repo = "search", branch = "master")
}
\arguments{
\item{account}{accountname or search for}

\item{repo}{repositoryname or search for}

\item{branch}{master, develop etc.}
}
\value{
markdown text to be placed in readme
}
\description{
Give or search for github name and repo.
}
