% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{licbadgebuilder}
\alias{licbadgebuilder}
\title{Add licence badge}
\usage{
licbadgebuilder(licencetype)
}
\arguments{
\item{licencetype}{one of GPL-3, GPL-2, MIT, or CC0.}
}
\value{
markdown
}
\description{
Add licence badge
}
\examples{
licbadgebuilder("GPL-3")
}
\seealso{
Other badges: \code{\link{codecovbadge}},
  \code{\link{cranbadge}},
  \code{\link{last_change_badge_static}},
  \code{\link{last_change_badge}},
  \code{\link{minimal_r_version_badge}},
  \code{\link{packageversionbadge}},
  \code{\link{projectstatusbadge}},
  \code{\link{travisbadge}}
}
