% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{licbadgebuilder}
\alias{licbadgebuilder}
\title{Build badges for licences}
\usage{
licbadgebuilder(licencetype)
}
\arguments{
\item{licencetype}{one of GPL-3, GPL-2, MIT, or CC0.}
}
\value{
markdown
}
\description{
Build badges for licences
}

