% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{licbadgebuilder}
\alias{licbadgebuilder}
\title{Add licence badge}
\usage{
licbadgebuilder(licencetype)
}
\arguments{
\item{licencetype}{one of GPL-3, GPL-2, MIT, or CC0.}
}
\value{
markdown
}
\description{
Add licence badge
}
\examples{
badgereader:::licbadgebuilder("GPL-3")
}
\seealso{
Other badges: \code{\link{badge_codecov}},
  \code{\link{badge_cran}},
  \code{\link{badge_github_fork}},
  \code{\link{badge_github_star}},
  \code{\link{badge_last_change_static}},
  \code{\link{badge_last_change}},
  \code{\link{badge_licence}},
  \code{\link{badge_minimal_r_version}},
  \code{\link{badge_packageversion}},
  \code{\link{badge_projectstatus}},
  \code{\link{badge_rdocumentation}},
  \code{\link{badge_travis}}
}
