% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{licbadgebuilder}
\alias{licbadgebuilder}
\title{Add licence badge}
\usage{
licbadgebuilder(licencetype)
}
\arguments{
\item{licencetype}{one of GPL-3, GPL-2, MIT, or CC0.}
}
\value{
markdown
}
\description{
Add licence badge
}
\examples{
badgecreatr:::licbadgebuilder("GPL-3")
}
