% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{licencepaste}
\alias{licencepaste}
\title{return markdown with image and link.}
\usage{
licencepaste(imagelink, referlink, name = "Licence")
}
\arguments{
\item{imagelink}{link to image file}

\item{referlink}{link to where to send to on click}

\item{name}{what to call the button}
}
\value{
markdown
}
\description{
for internal use. used within licencebuilder
}

