% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{packageversionbadge}
\alias{packageversionbadge}
\title{Place a badge with the version of your package.}
\usage{
packageversionbadge()
}
\value{
markdown to put into readme.rmd
}
\description{
Place a badge with the version of your package which is automatically read from
your description file.
}
\examples{
packageversionbadge()
}
