% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{projectstatusbadge}
\alias{projectstatusbadge}
\title{Add project status badge}
\usage{
projectstatusbadge(status)
}
\arguments{
\item{status}{one of concept, wip, suspended, abandoned, active, inactive or unsupported}
}
\value{
text to put into rmd
}
\description{
Add project status badge
}
\examples{
 
projectstatusbadge("unsupported")
}
