% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_placers.R
\name{travisbadge}
\alias{travisbadge}
\title{Travisbadge creates travis badge.}
\usage{
travisbadge(ghaccount, ghrepo, branch)
}
\arguments{
\item{ghaccount}{githubaccountname}

\item{ghrepo}{githubrepositoryname}

\item{branch}{master, develop etc}
}
\value{
link to travis image
}
\description{
Travisbadge creates travis badge.
}
\examples{
travisbadge(ghaccount = "johntest", ghrepo = "yourreponame", branch = "master")
}
